Store guild member level as uint

+++ src/server/game/Guilds/Guild.h
@@ -288,13 +288,13 @@ class TC_GAME_API Guild
                 Member(ObjectGuid::LowType guildId, ObjectGuid guid, uint8 rankId);
 
                 void SetStats(Player* player);
-                void SetStats(std::string const& name, uint8 level, uint8 _class, uint8 gender, uint32 zoneId, uint32 accountId);
+                void SetStats(std::string const& name, uint32 level, uint8 _class, uint8 gender, uint32 zoneId, uint32 accountId);
                 bool CheckStats() const;
 
                 void SetPublicNote(std::string const& publicNote);
                 void SetOfficerNote(std::string const& officerNote);
                 void SetZoneID(uint32 id) { m_zoneId = id; }
-                void SetLevel(uint8 var) { m_level = var; }
+                void SetLevel(uint32 var) { m_level = var; }
 
                 void AddFlag(uint8 var) { m_flags |= var; }
                 void RemFlag(uint8 var) { m_flags &= ~var; }
@@ -311,7 +311,7 @@ class TC_GAME_API Guild
                 std::string GetPublicNote() const { return m_publicNote; }
                 std::string GetOfficerNote() const { return m_officerNote; }
                 uint8 GetClass() const { return m_class; }
-                uint8 GetLevel() const { return m_level; }
+                uint32 GetLevel() const { return m_level; }
                 uint8 GetGender() const { return m_gender; }
                 uint8 GetFlags() const { return m_flags; }
                 uint32 GetZoneId() const { return m_zoneId; }
@@ -337,7 +337,7 @@ class TC_GAME_API Guild
                 ObjectGuid m_guid;
                 std::string m_name;
                 uint32 m_zoneId;
-                uint8 m_level;
+                uint32 m_level;
                 uint8 m_class;
                 uint8 m_gender;
                 uint8 m_flags;
